vllm-proxy/pyproject.toml
[tool.poetry]
name = "vllm-proxy"
version = "0.3.0"
description = "BRC-100 TEE inference proxy for Mandala Network"
authors = ["Shawn TIAN", "Mandala Network"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.6"}
uvicorn = {extras = ["standard"], version = "^0.34.0"}
cachetools = "^5.5.0"
dstack-sdk = "^0.5.0"
redis = "^5.2.1"
nv-ppcie-verifier = "^1.5.0"
bsv-sdk = ">=2.0.0b1"
bsv-wallet-toolbox = ">=2.0.1"
SQLAlchemy = ">=2.0"
requests = "^2.31.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-asyncio = "^0.25.3"
respx = "^0.22.0"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
